Phase 2 Step 1: アドレス型を int から long へ昇格
Memory/Segment/Elf/Cpu/Process の全アドレスフィールドと API シグネチャを
int → long に変更。32bit ELF では上位 32bit が 0 のまま流れるため挙動は変わらない。

主な変更:
- AllocInfo.address, Memory.mark_address/cache_address/memory_top → long
- Memory: load8/store8/load16..store64/fetch/storeString/loadString/dump/in/alloc/free → long address
- Segment: p_vaddr, p_paddr → long; in/peekb/peekbs/pokeb/expand_memory/segment_end → long
- Elf: e_entry, brk, symadrs → long; get_entry/get_curbrk/set_curbrk/get_symbol → long
- Cpu: ip, next_ip → long; ea() → long; get_ip/set_ip/get_sp/set_sp/fetch/disasm_str → long
- XDecoder: decode/cache_check → long ip
- Process: ip, handler_hook, sig_no_embed_adrs, handler_embed_adrs → long
- RootSysinfo: stack_bottom → long
- Util: hexstr(long, int) オーバーロード追加
- Operand: operand_str(long address)
- Syscall: sys_mmap の adrs → long, sys_brk でキャスト追加

Phase 0 回帰テスト 5/5 PASS 確認済み。

+++ src/main/java/emulin/Operand.java
@@ -89,7 +89,7 @@ public class Operand
   }
 
   // オペランド文字列を返す
-  public String operand_str( int address ) {
+  public String operand_str( long address ) {
     String sbuf = "";
     if( kind == REG ) {
       return( reg_name( reg_no ));
@@ -109,7 +109,7 @@ public class Operand
       return( sbuf );
     }
     if( kind == DISP ) {
-      return( "0x" + Util.hexstr( address + disp, 8 ));
+      return( "0x" + Util.hexstr( address + (long)disp, 8 ));
     }
     if( kind == IMM ) {
       return( "0x" + Util.hexstr( imm, 0 ));
